in the process of checking/fixing opcodes using the test suite from https://github.com/Klaus2m5/6502_65C02_functional_tests/

+++ emushared/include/IMemory.h
@@ -19,14 +19,14 @@ public:
      * @param b on successful return, the byte read
      * @return 0 on success, or errno
      */
-    virtual uint8_t readByte(uint16_t address, uint8_t* b) = 0;
+    virtual uint8_t readByte(uint32_t address, uint8_t* b) = 0;
     /**
      * read word (2 bytes) at address
      * @param address the address in the emulated memory
      * @param w on successful return, the word read
      * @return 0 on success, or errno
      */
-    virtual uint16_t readWord(uint16_t address, uint16_t* w) = 0;
+    virtual uint16_t readWord(uint32_t address, uint16_t* w) = 0;
 
     /**
      * write byte at address
@@ -34,7 +34,7 @@ public:
      * @param b byte to be written
      * @return 0 on success, or errno
      */
-    virtual int writeByte(uint16_t address, uint8_t b) = 0;
+    virtual int writeByte(uint32_t address, uint8_t b) = 0;
 
     /**
      * write word (2 bytes) at address
@@ -42,7 +42,7 @@ public:
      * @param w the word to be written (big/little endian is handled internally)
      * @return 0 on success, or errno
      */
-    virtual int writeWord(uint16_t address, uint16_t w) = 0;
+    virtual int writeWord(uint32_t address, uint16_t w) = 0;
 
     /**
      * write buffer at address
@@ -51,7 +51,7 @@ public:
      * @param size size of buffer to be written
      * @return 0 on success, or errno
      */
-    virtual int writeBytes(uint16_t address, uint8_t* b, uint16_t size) = 0;
+    virtual int writeBytes(uint32_t address, uint8_t* b, uint32_t size) = 0;
 
     /**
      * read buffer at address
@@ -61,14 +61,13 @@ public:
      * @param readSize size of read
      * @return 0 on success, or errno
      */
-    virtual int readBytes(uint16_t address, uint8_t* b, uint16_t bufferSize, uint16_t readSize) = 0;
+    virtual int readBytes(uint32_t address, uint8_t* b, uint32_t bufferSize, uint32_t readSize) = 0;
 
     /**
      * initializes emulated memory
      * @return 0 on success, or errno
      */
     virtual int init() = 0;
-
 };
 
 #endif //VC64_EMU_IMEMORY_H
